changes in flight search request

+++ src/main/java/com/brownfield/app/entity/Inventory.java
@@ -19,5 +19,5 @@ public class Inventory {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private long inventoryId;
-    private long count;
+    private Integer count;
 }
